requirements.txt
streamlit==1.34.0
openai-whisper==20231117
numpy==1.26.4
haystack-ai==2.1.2
ollama-haystack==0.0.6
python-dotenv==1.0.1
ffmpeg==1.4
streamlit-audiorecorder==0.0.5
gTTS==2.5.1
